tests: trim cross-file redundancy after audit
An audit of overlapping tests found 12 edge_cases tests covering the
same assertions as error_handling, chunk_boundaries, and corruption,
plus a couple of orphans that lived in the wrong file and the entire
cross_validation suite which duplicated count assertions already
covered by missing_message_types (real-fixture reference checks) and
real_world_fixtures (smoke tests).

Changes:

  tests/cross_validation.rs: deleted entirely. Its assertions (count
    totals per topic, parameter totals, GPS reference values, info
    message contents) are covered by:
      - missing_message_types.rs: info_sample_ulg_matches_pyulog_reference,
        dropout_sample_ulg_matches_pyulog_reference
      - real_world_fixtures.rs: message-count and topic-presence asserts
        against the same fixtures

  tests/edge_cases.rs: cut 12 tests that duplicated assertions in other
    files. Kept the unique ones covering message-sequencing violations
    (data before flag_bits), size-boundary edges (zero-length message,
    max-length header), positive tests for deliberate parser behavior
    (DATA_APPENDED flag accepted, unknown type silently skipped), and
    the empty/not-a-log-file fixture path. Cuts:
      - test_header_plus_one_byte, test_header_plus_two_bytes,
        test_truncated_message_body (all subsumed by chunk_boundaries
        which exercises reassembly, not just "doesn't error")
      - test_wrong_magic_bytes, test_almost_valid_magic,
        test_random_garbage_file (same assertions as test_all_zeros_file
        or corruption::header_with_* variants)
      - test_data_message_unregistered_msg_id, test_data_message_wrong_size
        (covered by error_handling and corruption)
      - test_corrupt_format_no_colon, test_corrupt_format_empty_name
        (covered by error_handling::malformed_format_strings_are_rejected)
      - test_format_duplicate_field_name,
        test_flag_bits_incompatible_flags_reject (covered by their
        error_handling counterparts)

  Other trims following the same audit:
    - chunk_boundaries::test_sample_ulg_chunk_sizes and
      test_random_chunk_sizes: subsumed by test_chunk_sizes_sweep
    - corruption::unknown_first_message_type_does_not_panic:
      subsumed by edge_cases::test_unknown_message_type_ignored
    - error_handling::error_works_with_question_mark_operator and
      undersized_flag_bits_message_is_rejected: redundant with other
      tests in the same file and in edge_cases
    - mmap_backend::test_all_fixtures_byte_slice_matches_file_parse:
      parametrised sweep over fixtures that are already each tested
    - regression_bugs::test_off_by_one_with_trailing_byte: trivial
      variant of test_off_by_one_exact_buffer_boundary
    - unified_parser_api::stream_parser_rejects_empty_input and
      stream_parser_byte_count_matches_one_shot_vs_chunked: trivial
      behaviour / covered by chunk_boundaries

  Orphan relocated: mixed_stream_surfaces_every_variant_and_preserves_data
    moved from missing_message_types to unified_parser_api. It's a
    cross-cutting integration test that doesn't belong next to per-type
    smoke tests.

Net: 175 -> 141 #[test] fns (-34), 4636 -> 4009 LOC (-627 lines),
17 -> 16 test files. Nothing new left uncovered; cargo test + clippy
-D warnings + fmt all clean.
