+++ tests/edge_cases.rs
@@ -1,5 +1,13 @@
-//! Edge case and corruption tests. Verify the parser does not panic or hang
-//! on malformed input.
+//! Structural edge cases: inputs that are technically malformed but not
+//! produced by random corruption. Covers empty/minimal streams, size
+//! boundaries on the message framing, message sequencing violations
+//! (data before definitions, out-of-order FlagBits), and positive tests
+//! for behavior the parser deliberately allows (DATA_APPENDED flag,
+//! unknown message types).
+//!
+//! Corruption-fuzz mutations live in tests/corruption.rs. Specific
+//! error-message content assertions live in tests/error_handling.rs.
+//! Chunk-boundary reassembly lives in tests/chunk_boundaries.rs.
 
 mod helpers;
 
@@ -24,52 +32,9 @@ fn test_empty_file() {
     assert!(result.is_ok(), "Empty input should not error");
 }
 
-#[test]
-fn test_only_header_no_messages() {
-    let builder = ULogBuilder::new();
-    let bytes = builder.build();
-    assert_eq!(bytes.len(), 16);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_ok(), "Header-only input should not error");
-}
-
-#[test]
-fn test_header_plus_one_byte() {
-    let mut bytes = ULogBuilder::new().build();
-    bytes.push(0x00); // partial message header
-    let result = parse_bytes(&bytes);
-    assert!(
-        result.is_ok(),
-        "Partial message header should stash as leftover, not error"
-    );
-}
-
-#[test]
-fn test_header_plus_two_bytes() {
-    let mut bytes = ULogBuilder::new().build();
-    bytes.extend_from_slice(&[0x00, 0x00]); // 2 of 3 message header bytes
-    let result = parse_bytes(&bytes);
-    assert!(result.is_ok(), "Incomplete message header should not error");
-}
-
 // =============================================================================
-// Truncated messages
+// Pathological message-size boundaries
 // =============================================================================
-
-#[test]
-fn test_truncated_message_body() {
-    let mut bytes = ULogBuilder::new().build();
-    // Message header claiming 100 bytes, but only 50 bytes of body
-    bytes.extend_from_slice(&100u16.to_le_bytes());
-    bytes.push(b'B'); // FlagBits type
-    bytes.extend_from_slice(&[0u8; 50]); // only 50 of 100 bytes
-    let result = parse_bytes(&bytes);
-    // Should stash in leftover, not error (waiting for more data)
-    assert!(
-        result.is_ok(),
-        "Truncated message body should be stashed as leftover"
-    );
-}
 
 #[test]
 fn test_zero_length_message() {
@@ -100,43 +65,14 @@ fn test_max_length_message_header() {
 }
 
 // =============================================================================
-// Invalid headers
+// Not even remotely a ULog
 // =============================================================================
-
-#[test]
-fn test_wrong_magic_bytes() {
-    let bytes = vec![0x00; 16]; // all zeros, wrong magic
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Wrong magic should error");
-}
 
 #[test]
 fn test_all_zeros_file() {
     let bytes = vec![0u8; 1024];
     let result = parse_bytes(&bytes);
     assert!(result.is_err(), "All-zero file should error on header");
-}
-
-#[test]
-fn test_random_garbage_file() {
-    // Fixed seed for reproducibility
-    let mut bytes = Vec::with_capacity(1024);
-    let mut state: u32 = 0xDEADBEEF;
-    for _ in 0..1024 {
-        state = state.wrapping_mul(1103515245).wrapping_add(12345);
-        bytes.push((state >> 16) as u8);
-    }
-    let result = parse_bytes(&bytes);
-    // Should error on invalid header, not panic
-    assert!(result.is_err(), "Random garbage should error on header");
-}
-
-#[test]
-fn test_almost_valid_magic() {
-    let mut bytes = ULogBuilder::new().build();
-    bytes[6] = 0x00; // corrupt last magic byte
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Corrupted magic byte should error");
 }
 
 // =============================================================================
@@ -187,81 +123,9 @@ fn test_flag_bits_not_first_message() {
     }
 }
 
-#[test]
-fn test_data_message_unregistered_msg_id() {
-    let mut builder = ULogBuilder::new();
-    builder
-        .flag_bits()
-        .format("test_topic", &[("uint64_t", "timestamp"), ("float", "x")]);
-    // Add subscription for msg_id=0
-    builder.add_logged(0, 0, "test_topic");
-    // Data message for msg_id=99 (never registered)
-    let mut payload = Vec::new();
-    payload.extend_from_slice(&99u16.to_le_bytes());
-    payload.extend_from_slice(&1000u64.to_le_bytes());
-    payload.extend_from_slice(&1.0f32.to_le_bytes());
-    builder.data_raw(&payload);
-    let mut bytes = builder.build();
-    bytes.push(0x00);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Unregistered msg_id should error");
-}
-
-#[test]
-fn test_data_message_wrong_size() {
-    let mut builder = ULogBuilder::new();
-    builder
-        .flag_bits()
-        .format("test_topic", &[("uint64_t", "timestamp"), ("float", "x")])
-        .add_logged(0, 0, "test_topic");
-    // Data message with wrong payload size (too short)
-    let mut payload = Vec::new();
-    payload.extend_from_slice(&0u16.to_le_bytes()); // msg_id
-    payload.extend_from_slice(&1000u64.to_le_bytes()); // timestamp only, missing float
-    builder.data_raw(&payload);
-    let mut bytes = builder.build();
-    bytes.push(0x00);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Data message with wrong size should error");
-}
-
 // =============================================================================
-// Invalid format strings
+// Format-string edge cases not covered by error_handling's parameterized check
 // =============================================================================
-
-#[test]
-fn test_corrupt_format_no_colon() {
-    let mut builder = ULogBuilder::new();
-    builder.flag_bits();
-    // Raw format message without colon separator
-    builder.unknown_message(b'F', b"invalid_format_string");
-    let mut bytes = builder.build();
-    bytes.push(0x00);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Format without colon should error");
-}
-
-#[test]
-fn test_corrupt_format_empty_name() {
-    let mut builder = ULogBuilder::new();
-    builder.flag_bits();
-    builder.unknown_message(b'F', b":uint64_t timestamp");
-    let mut bytes = builder.build();
-    bytes.push(0x00);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Format with empty name should error");
-}
-
-#[test]
-fn test_format_duplicate_field_name() {
-    let mut builder = ULogBuilder::new();
-    builder.flag_bits();
-    builder.unknown_message(b'F', b"dup:uint64_t x;float x");
-    let mut bytes = builder.build();
-    bytes.push(0x00);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Format with duplicate field should error");
-}
 
 #[test]
 fn test_format_empty_fields() {
@@ -278,19 +142,6 @@ fn test_format_empty_fields() {
 // =============================================================================
 // Flag bits edge cases
 // =============================================================================
-
-#[test]
-fn test_flag_bits_incompatible_flags_reject() {
-    let mut builder = ULogBuilder::new();
-    // Set unknown incompat flag in byte 1
-    let mut incompat = [0u8; 8];
-    incompat[1] = 0xFF;
-    builder.flag_bits_with_incompat(&incompat);
-    let mut bytes = builder.build();
-    bytes.push(0x00);
-    let result = parse_bytes(&bytes);
-    assert!(result.is_err(), "Unknown incompat flags should error");
-}
 
 #[test]
 fn test_flag_bits_data_appended_flag_accepted() {
@@ -322,7 +173,7 @@ fn test_flag_bits_too_short() {
 }
 
 // =============================================================================
-// Unknown message types
+// Unknown message types are silently skipped (positive test)
 // =============================================================================
 
 #[test]
@@ -341,7 +192,7 @@ fn test_unknown_message_type_ignored() {
 }
 
 // =============================================================================
-// Not a log file
+// Non-ULog input file
 // =============================================================================
 
 #[test]
